+++ src/directory.rs
@@ -1,4 +1,4 @@
-use std::path::Path;
+use std::{path::Path, time::Duration};
 
 use notify::{Event, EventKind, RecursiveMode, Watcher};
 use tokio::sync::watch::{self, Receiver};
@@ -6,6 +6,7 @@ use tokio::sync::watch::{self, Receiver};
 // Listen to events in a directory, and send relevant updates in a watch channel.
 pub(crate) fn watch_directory<W: Watcher>(
     directory: &Path,
+    poll_interval: Duration,
 ) -> color_eyre::Result<(W, Receiver<()>)> {
     let (tx, rx) = watch::channel(());
     let mut watcher = W::new(
@@ -19,7 +20,9 @@ pub(crate) fn watch_directory<W: Watcher>(
                 }
             };
         },
-        notify::Config::default(),
+        notify::Config::default()
+            .with_follow_symlinks(true)
+            .with_poll_interval(poll_interval),
     )?;
     watcher.watch(directory, RecursiveMode::Recursive)?;
     Ok((watcher, rx))