+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 ## Unreleased
 
+### Added
+
+- Add `--directory-poll-interval` CLI flag.
+
 ### Fixed
 
-- Proper debouncing for certificates/keys file system watchers.
+- Don't unwrap ACME challenge TLS stream acceptor.
+- Switch directory watcher strategy in order to follow symlinks.
 
 ## 0.8.6 (2025-12-22)